# docs/swagger.yaml
basePath: /
definitions:
  models.Car:
    properties:
      id:
        example: 1
        type: integer
      name:
        example: Toyota Land Cruiser 100
        type: string
      owner_id:
        example: 1
        type: integer
    type: object
  models.Equipment:
    properties:
      expedition_id:
        example: 1
        type: integer
      id:
        example: 1
        type: integer
      name:
        example: GPS Navigator
        type: string
    type: object
  models.Good:
    properties:
      expedition_id:
        example: 1
        type: integer
      id:
        example: 1
        type: integer
      name:
        example: First Aid Kit
        type: string
    type: object
...
host: localhost:8081
info:
  contact: {}
  title: Travel Planner
  version: "1.0"
paths:
  /cars:
    delete:
      description: Удаление автомобиля по идентификатору
      parameters:
      - description: Идентификатор автомобиля
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: Автомобиль успешно удален
        "400":
          description: Некорректный ID
          schema:
            type: string
      summary: Удалить автомобиль
      tags:
      - Автомобили
    get:
      description: Получить список всех автомобилей или конкретный автомобиль по ID
      parameters:
      - description: Идентификатор автомобиля (необязательно)
        in: query
        name: id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Успешное получение автомобиля
          schema:
            $ref: '#/definitions/models.Car'
        "400":
          description: Некорректный запрос
          schema:
            type: string
        "404":
          description: Автомобиль не найден
          schema:
            type: string
      summary: Получить автомобиль(и)
      tags:
      - Автомобили
    patch:
      consumes:
      - application/json
      description: Частичное обновление данных автомобиля
      parameters:
      - description: Идентификатор автомобиля
        in: query
        name: id
        required: true
        type: integer
      - description: Данные для обновления
        in: body
        name: car
        required: true
        schema:
          $ref: '#/definitions/models.Car'
      produces:
      - application/json
      responses:
        "200":
          description: Автомобиль успешно обновлен
          schema:
            $ref: '#/definitions/models.Car'
        "400":
          description: Некорректный ID или данные
          schema:
            type: string
        "404":
          description: Автомобиль не найден
          schema:
            type: string
      summary: Обновить автомобиль
      tags:
      - Автомобили
    post:
      consumes:
      - application/json
      description: Добавить новый автомобиль в базу данных
      parameters:
      - description: Данные автомобиля
        in: body
        name: car
        required: true
        schema:
          $ref: '#/definitions/models.Car'
      produces:
      - application/json
      responses:
        "201":
          description: Автомобиль успешно создан
          schema:
            $ref: '#/definitions/models.Car'
        "400":
          description: Некорректные входные данные
          schema:
            type: string
        "409":
          description: Конфликт данных
          schema:
            type: string
      summary: Создать автомобиль
      tags:
      - Автомобили
  /equip:
    delete:
      description: Удалить снаряжение по ID
      parameters:
      - description: ID снаряжения
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: Снаряжение удалено
        "400":
          description: Неверный ID
          schema:
            type: string
        "409":
          description: Ошибка удаления
          schema:
            type: string
      summary: Удалить снаряжение
      tags:
      - Снаряжение
    get:
      description: Получить список всего снаряжения или конкретный экземпляр по ID
      parameters:
      - description: ID снаряжения
        in: query
        name: id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Список снаряжения
          schema:
            items:
              $ref: '#/definitions/models.Equipment'
            type: array
        "400":
          description: Неверный ID
          schema:
            type: string
        "404":
          description: Снаряжение не найдено
          schema:
            type: string
      summary: Получить снаряжение
      tags:
      - Снаряжение
    patch:
      consumes:
      - application/json
      description: Частичное обновление данных снаряжения
      parameters:
      - description: ID снаряжения
        in: query
        name: id
        required: true
        type: integer
      - description: Обновляемые поля
        in: body
        name: equipment
        required: true
        schema:
          $ref: '#/definitions/models.Equipment'
      produces:
      - application/json
      responses:
        "200":
          description: Обновленное снаряжение
          schema:
            $ref: '#/definitions/models.Equipment'
        "400":
          description: Неверные данные
          schema:
            type: string
        "404":
          description: Снаряжение не найдено
          schema:
            type: string
        "409":
          description: Ошибка обновления
          schema:
            type: string
      summary: Обновить снаряжение
      tags:
      - Снаряжение
    post:
      consumes:
      - application/json
      description: Создать новое снаряжение
      parameters:
      - description: Данные снаряжения
        in: body
        name: equipment
        required: true
        schema:
          $ref: '#/definitions/models.Equipment'
      produces:
      - application/json
      responses:
        "201":
          description: Созданное снаряжение
          schema:
            $ref: '#/definitions/models.Equipment'
        "400":
          description: Ошибка валидации
          schema:
            type: string
        "409":
          description: Конфликт данных
          schema:
            type: string
      summary: Создать снаряжение
      tags:
      - Снаряжение
  /goods:
    delete:
      description: Удаление позиции продукта по ID
      parameters:
      - description: ID продукта
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: Успешное удаление
        "400":
          description: Некорректный ID
          schema:
            type: string
        "404":
          description: Не найдено
          schema:
            type: string
        "500":
          description: Ошибка сервера
          schema:
            type: string
      summary: Удалить позицию
      tags:
      - Продукты
    get:
      description: Получение списка всех продуктов или конкретной позиции по ID
      parameters:
      - description: ID продукта
        in: query
        name: id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Конкретная позиция
          schema:
            $ref: '#/definitions/models.Good'
        "400":
          description: Некорректный запрос
          schema:
            type: string
        "404":
          description: Не найдено
          schema:
            type: string
        "500":
          description: Ошибка сервера
          schema:
            type: string
      summary: Получить список продуктов
      tags:
      - Продукты
    patch:
      consumes:
      - application/json
      description: Обновление существующей позиции (частичное обновление)
      parameters:
      - description: ID продукта
        in: query
        name: id
        required: true
        type: integer
      - description: Данные для обновления
        in: body
        name: good
        required: true
        schema:
          $ref: '#/definitions/models.Good'
      produces:
      - application/json
      responses:
        "200":
          description: Обновленная позиция
          schema:
            $ref: '#/definitions/models.Good'
        "400":
          description: Некорректный запрос
          schema:
            type: string
        "404":
          description: Не найдено
          schema:
            type: string
        "409":
          description: Конфликт данных
          schema:
            type: string
      summary: Обновить позицию
      tags:
      - Продукты
    post:
      consumes:
      - application/json
      description: Создание новой позиции продуктов
      parameters:
      - description: Данные продукта
        in: body
        name: good
        required: true
        schema:
          $ref: '#/definitions/models.Good'
      produces:
      - application/json
      responses:
        "201":
          description: Созданная позиция
          schema:
            $ref: '#/definitions/models.Good'
        "400":
          description: Некорректные данные
          schema:
            type: string
        "409":
          description: Конфликт - позиция уже существует
          schema:
            type: string
      summary: Создать новую позицию
      tags:
      - Продукты
...
swagger: "2.0"